requirements.txt
flask
line-bot-sdk
google-generativeai
chromadb
crawl4ai
playwright
nest_asyncio
gunicorn